OkayCMS 2.2.0

+++ api/Pages.php
@@ -4,6 +4,9 @@ require_once('Okay.php');
 
 class Pages extends Okay {
 
+    // Системные url
+    public $system_pages = array('catalog', 'products', 'all-products', 'discounted', 'bestsellers', 'brands', 'blog', 'news', 'wishlist', 'comparison', 'cart', 'order', 'contact', 'user', '404', '');
+    
     /*Выборка конкрентной страницы*/
     public function get_page($id) {
         if(gettype($id) == 'string') {
@@ -17,7 +20,6 @@ class Pages extends Okay {
         $query = "SELECT 
                 p.id, 
                 p.url, 
-                p.menu_id, 
                 p.position, 
                 p.visible, 
                 p.last_modify, 
@@ -36,13 +38,8 @@ class Pages extends Okay {
 
     /*Выборка всех страниц*/
     public function get_pages($filter = array()) {
-        $menu_filter = '';
         $visible_filter = '';
         $pages = array();
-        
-        if(isset($filter['menu_id'])) {
-            $menu_filter = $this->db->placehold('AND p.menu_id in (?@)', (array)$filter['menu_id']);
-        }
         
         if(isset($filter['visible'])) {
             $visible_filter = $this->db->placehold('AND p.visible = ?', intval($filter['visible']));
@@ -52,7 +49,6 @@ class Pages extends Okay {
         $query = "SELECT 
                 p.id, 
                 p.url, 
-                p.menu_id, 
                 p.position, 
                 p.visible, 
                 p.last_modify, 
@@ -61,7 +57,6 @@ class Pages extends Okay {
             $lang_sql->join 
             WHERE 
                 1 
-                $menu_filter 
                 $visible_filter 
             ORDER BY p.position
         ";
@@ -118,6 +113,11 @@ class Pages extends Okay {
     /*Удаление страницы*/
     public function delete_page($id) {
         if(!empty($id)) {
+            // Запретим удаление системных ссылок
+            $page = $this->get_page(intval($id));
+            if (in_array($page->url, $this->system_pages)) {
+                return false;
+            }
             $query = $this->db->placehold("DELETE FROM __pages WHERE id=? LIMIT 1", intval($id));
             if($this->db->query($query)) {
                 $this->db->query("DELETE FROM __lang_pages WHERE page_id=?", intval($id));
@@ -127,22 +127,4 @@ class Pages extends Okay {
         return false;
     }
 
-    /*Выборка всех меню*/
-    public function get_menus() {
-        $menus = array();
-        $query = "SELECT * FROM __menu ORDER BY position";
-        $this->db->query($query);
-        foreach($this->db->results() as $menu) {
-            $menus[$menu->id] = $menu;
-        }
-        return $menus;
-    }
-
-    /*Выборка конкретного меню*/
-    public function get_menu($menu_id) {
-        $query = $this->db->placehold("SELECT * FROM __menu WHERE id=? LIMIT 1", intval($menu_id));
-        $this->db->query($query);
-        return $this->db->result();
-    }
-    
 }
